pyproject.toml
[project]
name = "time-sheets"
version = "0.1.0"
description = ""
authors = [
    {name = "Dan Ballance",email = "work@danballance.uk"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest-cov (>=6.0.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
